Update vviki.vim
This plugin is really great. I have only one issue with it: it uses the `link:url[text]` syntax for references. Although this is fine for external Links, it does not work for cross-references with AsciiDoc documents. The documentation of AsciiDoctor (https://asciidoctor.org/docs/user-manual/#inter-document-cross-references) recommends here using the "<<url#,text>>" syntax.

This commit tries to fix this, it seems to work well ;-).